## Description

In #8694 cache for `AccAddress.String()` was added as a simple map without control and estimates how that map will grow. This leads to unmanageable memory growth, which could put a node under a risk.

In this PR, 
+ we add a benchmark for the Address.String method.
+ we add an LRU cache instead of the map. 
+ remove deferred calls
+ optimize the `Address.Empty` function

Before:
```
BenchmarkAccAddressString-4   	 4533326	       268 ns/op	     144 B/op	       3 allocs/op
BenchmarkAccAddressString-4   	 4592979	       283 ns/op	     144 B/op	       3 allocs/op
```
After:
```
BenchmarkAccAddressString-4   	 4180477	       279 ns/op	     144 B/op	       3 allocs/op
BenchmarkAccAddressString-4   	 4633738	       274 ns/op	     144 B/op	       3 allocs/op
```

---

Before we can merge this PR, please make sure that all the following items have been
checked off. If any of the checklist items are not applicable, please leave them but
write a little note why.

- [x] Targeted PR against correct branch (see [CONTRIBUTING.md](https://github.com/cosmos/cosmos-sdk/blob/master/CONTRIBUTING.md#pr-targeting))
- [ ] Linked to Github issue with discussion and accepted design OR link to spec that describes this work.
- [x] Code follows the [module structure standards](https://github.com/cosmos/cosmos-sdk/blob/master/docs/building-modules/structure.md).
- [x] Wrote unit and integration [tests](https://github.com/cosmos/cosmos-sdk/blob/master/CONTRIBUTING.md#testing)
- [ ] Updated relevant documentation (`docs/`) or specification (`x/<module>/spec/`)
- [ ] Added relevant `godoc` [comments](https://blog.golang.org/godoc-documenting-go-code).
- [ ] Added a relevant changelog entry to the `Unreleased` section in `CHANGELOG.md`
- [x] Re-reviewed `Files changed` in the Github PR explorer
- [ ] Review `Codecov Report` in the comment section below once CI passes