@robert-zaremba 2021-02-26
BTW: we could add same thing for `ValAddress` and `ConsAddress`.
However, to be honest, I don't see much sense adding cache for `Address`. There are tonnes of other, better places for optimization without using a cache. Since it was added, I wanted to fix the memory issue.