@robert-zaremba 2021-03-01
We can make more sophisticated benchmarks later (feel free to create a task for that). 
I will add the unsafe conversion - but I don't think this is needed (especially if we want to use `defer` - which will create a new function stack)