@robert-zaremba 2021-03-01
1. Updated cache size
2. Moved Empty check before checking cache. Now Empty check is  very fast - so we don't need to cache it.
3. Applied same mechanism to other address types: `ValAddress` and `ConsAddress`.
4. Used unsafe bytes->string conversion.