@robert-zaremba 2021-03-02
@odeke-em  - adding more benchmarks is going out of scope and we should have a good to have a real world scenario here. We can do it in other issue. 

The intention of this PR was to remove the memory leak. I think it's better to have this update then rolling back the cache. I suggest to create a new issue with further benchmarks.

I added one more, non trivial address to the benchmark.