@robert-zaremba 2021-03-03
@odeke-em  your lock on this PR is still active. Please comment if there is something wrong in this PR. Adding more benchmarks is, imho, out of scope - we would need to go into use-cases  / scenario bench-marking to bring bore valuable information.