@robert-zaremba 2021-03-03
>  If the cost of using an LRU is worse that’s a problem.

The benchmark (which extensively uses cache) shows that's not worse. Moreover the new code uses less allocations - this is due to avoiding `defer` and optimizing "empty" case.