@robert-zaremba 2021-03-03
@colin-axner  What exactly and Where do you see the test failing?
Here, in the merge checks nothing was failing.
When running `make test-race` I see the following things failing:

+ FAIL: TestListSnapshots (137.46s)
+ FAIL	github.com/cosmos/cosmos-sdk/baseapp	468.134s

It's hard to digest the output of `maket test-race`. There are tonnes of lines related to `leveldb`....

The only interesting thing I found is the following warning:
```
WARNING: DATA RACE
Write at 0x00c000fbf470 by goroutine 43:
  container/list.(*List).move()
      /usr/lib/go-1.15/src/container/list/list.go:123 +0x28a
  container/list.(*List).MoveToFront()
      /usr/lib/go-1.15/src/container/list/list.go:188 +0x23f
  github.com/hashicorp/golang-lru/simplelru.(*LRU).Get()
      /home/robert/go/pkg/mod/github.com/hashicorp/golang-lru@v0.5.4/simplelru/lru.go:75 +0xed
  github.com/cosmos/cosmos-sdk/types.AccAddress.String()
      /home/robert/go/src/github.com/cosmos/cosmos-sdk/types/address.go:271 +0x152
```
This access is OK - because we are fine with a concurrent read, even if it's done during the read - we write data only once, and it never depends on what we read. I can add a mutex there if we want to mute that warning. What do you think?