@colin-axner 2021-03-03
Ah, I got a little confused by not looking at the full logs

I also see this (and similar errors in a couple other places):
```
2021-03-03T03:52:25.9594698Z --- FAIL: TestIntegrationTestSuite (102.02s)
2021-03-03T03:52:25.9595884Z     --- FAIL: TestIntegrationTestSuite/TestCLIMultisign (11.95s)
2021-03-03T03:52:25.9597152Z     --- FAIL: TestIntegrationTestSuite/TestCLISignBatch (4.64s)
2021-03-03T03:52:25.9598103Z     testing.go:1038: race detected during execution of test
2021-03-03T03:52:25.9598655Z FAIL
2021-03-03T03:52:25.9599405Z FAIL	github.com/cosmos/cosmos-sdk/x/auth/client/cli	102.710s
```

My main concern is having the race-detector become stuck on always failing. I never check the logs, so I imagine it could be dangerous to have a test which always fails since if someone does introduce a race condition, it could go unnoticed

I have little context on this pr, so will differ to the decisions of others here. Maybe the other `test-race` tests are a good enough to ensure race conditions don't get introduced