@robert-zaremba 2021-03-03
Yeah, the logs are huuuuge. I will add a check and let's see what will happen.