@alessio 2021-03-03
Oh wow - why was mergify allowed to merge something whose tests are not all green in the first place?

@marbar3778 can we do something to prevent this from happening again?