@alessio 2021-03-03
I think I know why. The `race-detector-report` job was not mandatory for PRs to be merged in.