@alessio 2021-03-03
> It should be noted that I don't think any pr's can be merged until the race detector is fixed

The branch configuration is now fixed, so PRs will be blocked until the race is fixed. Does that work for you?