@robert-zaremba 2021-03-03
PR: https://github.com/cosmos/cosmos-sdk/pull/8767